Update Header and Footer.js and add a Home page
