+++ frontend/src/shared/components/Footer.js
@@ -1,12 +1,10 @@
-import React from "react";
-
 // Footer component for shared footer content
-const Footer = () => {
+function Footer() {
   return (
     <div>
       <p>&copy; 2025 TourismPulseNZ</p>
     </div>
   );
-};
+}
 
 export default Footer; // Exporting the Footer component for use in other parts of the application
